server/requirements.txt
requests
Django
Pillow
gunicorn
python-dotenv
ibm-watson>=7.0.0
ibm-cloud-sdk-core>=3.16.0
